requirements.txt
flask==2.3.0
requests==2.31.0
python-dotenv==1.0.0
openai==0.27.0
